Add tests to authentication branch
